+++ transformers/component_tests/flip/configured_transformer_test.go
@@ -1,5 +1,5 @@
 // VulcanizeDB
-// Copyright © 2018 Vulcanize
+// Copyright © 2019 Vulcanize
 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published by
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package flop
+package flip
 
 import (
 	"strconv"
@@ -24,7 +24,7 @@ import (
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	mcdStorage "github.com/makerdao/vdb-mcd-transformers/transformers/storage"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/flop"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/flip"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/test_helpers"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/test_data"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/event"
@@ -36,147 +36,110 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Executing the flop transformer", func() {
+var _ = Describe("Executing the flip transformer", func() {
 	var (
-		db                     = test_config.NewTestDB(test_config.NewTestNode())
-		flopperContractAddress = test_data.FlopAddress()
-		repository             = flop.FlopStorageRepository{ContractAddress: flopperContractAddress}
-		storageKeysLookup      = storage.NewKeysLookup(flop.NewKeysLoader(&mcdStorage.MakerStorageRepository{}, flopperContractAddress))
-		header                 = fakes.FakeHeader
-		transformer            storage.Transformer
-		headerID               int64
+		db                = test_config.NewTestDB(test_config.NewTestNode())
+		contractAddress   = test_data.EthFlipAddress()
+		repository        = flip.FlipStorageRepository{ContractAddress: contractAddress}
+		storageKeysLookup = storage.NewKeysLookup(flip.NewKeysLoader(&mcdStorage.MakerStorageRepository{}, contractAddress))
+		header            = fakes.FakeHeader
+		transformer       storage.Transformer
 	)
 
 	BeforeEach(func() {
 		test_config.CleanTestDB(db)
 		transformer = storage.Transformer{
-			HashedAddress:     vdbStorage.HexToKeccak256Hash(flopperContractAddress),
+			HashedAddress:     vdbStorage.HexToKeccak256Hash(contractAddress),
 			StorageKeysLookup: storageKeysLookup,
 			Repository:        &repository,
 		}
 		transformer.NewTransformer(db)
 		headerRepository := repositories.NewHeaderRepository(db)
 		var insertHeaderErr error
-		headerID, insertHeaderErr = headerRepository.CreateOrUpdateHeader(fakes.FakeHeader)
+		header.Id, insertHeaderErr = headerRepository.CreateOrUpdateHeader(header)
 		Expect(insertHeaderErr).NotTo(HaveOccurred())
-		header.Id = headerID
 	})
 
 	It("reads in a vat storage diff and persists it", func() {
-		vat := "0x1CC5ABe5C0464F3af2a10df0c711236a8446BF75"
 		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000002")
-		value := common.HexToHash("0000000000000000000000001cc5abe5c0464f3af2a10df0c711236a8446bf75")
+		value := common.HexToHash("000000000000000000000000284ecb5880cdc3362d979d07d162bf1d8488975d")
 		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
 
 		err := transformer.Execute(diff)
 		Expect(err).NotTo(HaveOccurred())
 
 		var vatResult test_helpers.VariableRes
-		err = db.Get(&vatResult, `SELECT diff_id, header_id, vat AS value from maker.flop_vat`)
+		err = db.Get(&vatResult,
+			`SELECT diff_id, header_id, vat AS value FROM maker.flip_vat`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(vatResult, diff.ID, headerID, vat)
+		test_helpers.AssertVariable(vatResult, diff.ID, header.Id, "0x284ecB5880CdC3362D979D07D162bf1d8488975D")
 	})
 
-	It("reads in a gem storage diff and persists it", func() {
-		gem := "0xAaF64BFCC32d0F15873a02163e7E500671a4ffcD"
+	It("reads in an ilk storage diff and persists it", func() {
+		ilk := "4554482d41000000000000000000000000000000000000000000000000000000"
 		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000003")
-		value := common.HexToHash("000000000000000000000000aaf64bfcc32d0f15873a02163e7e500671a4ffcd")
+		value := common.HexToHash(ilk)
 		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
+
 		err := transformer.Execute(diff)
 		Expect(err).NotTo(HaveOccurred())
 
-		var gemResult test_helpers.VariableRes
-		err = db.Get(&gemResult, `SELECT diff_id, header_id, gem AS value from maker.flop_gem`)
+		var ilkResult test_helpers.VariableRes
+		err = db.Get(&ilkResult,
+			`SELECT diff_id, header_id, ilk_id AS value FROM maker.flip_ilk`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(gemResult, diff.ID, headerID, gem)
+		ilkID, ilkErr := shared.GetOrCreateIlk(ilk, db)
+		Expect(ilkErr).NotTo(HaveOccurred())
+		test_helpers.AssertVariable(ilkResult, diff.ID, header.Id, strconv.FormatInt(ilkID, 10))
 	})
 
 	It("reads in a beg storage diff and persists it", func() {
-		beg := "1050000000000000000000000000"
 		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000004")
 		value := common.HexToHash("000000000000000000000000000000000000000003648a260e3486a65a000000")
 		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
+
 		err := transformer.Execute(diff)
 		Expect(err).NotTo(HaveOccurred())
 
 		var begResult test_helpers.VariableRes
-		err = db.Get(&begResult, `SELECT diff_id, header_id, beg AS value from maker.flop_beg`)
+		err = db.Get(&begResult,
+			`SELECT diff_id, header_id, beg AS value FROM maker.flip_beg`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(begResult, diff.ID, headerID, beg)
-	})
-
-	It("reads in a pad storage diff and persists it", func() {
-		pad := "1500000000000000000000000000"
-		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000005")
-		value := common.HexToHash("000000000000000000000000000000000000000004d8c55aefb8c05b5c000000")
-		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
-		err := transformer.Execute(diff)
-		Expect(err).NotTo(HaveOccurred())
-
-		var ttlResult test_helpers.VariableRes
-		err = db.Get(&ttlResult, `SELECT diff_id, header_id, pad AS value from maker.flop_pad`)
-		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(ttlResult, diff.ID, headerID, pad)
+		test_helpers.AssertVariable(begResult, diff.ID, header.Id, "1050000000000000000000000000")
 	})
 
 	It("reads in a ttl storage diff and persists it", func() {
-		ttl := "10800"
-		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000006")
+		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000005")
 		value := common.HexToHash("000000000000000000000000000000000000000000000002a300000000002a30")
 		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
+
 		err := transformer.Execute(diff)
 		Expect(err).NotTo(HaveOccurred())
 
 		var ttlResult test_helpers.VariableRes
-		err = db.Get(&ttlResult, `SELECT diff_id, header_id, ttl AS value from maker.flop_ttl`)
+		err = db.Get(&ttlResult,
+			`SELECT diff_id, header_id, ttl AS value FROM maker.flip_ttl`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(ttlResult, diff.ID, headerID, ttl)
+		test_helpers.AssertVariable(ttlResult, diff.ID, header.Id, "10800")
 	})
 
 	It("reads in a tau storage diff and persists it", func() {
-		ttl := "172800"
-		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000006")
+		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000005")
 		value := common.HexToHash("000000000000000000000000000000000000000000000002a300000000002a30")
 		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
-		err := transformer.Execute(diff)
-		Expect(err).NotTo(HaveOccurred())
-
-		var tauResult test_helpers.VariableRes
-		err = db.Get(&tauResult, `SELECT diff_id, header_id, tau AS value from maker.flop_tau`)
-		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(tauResult, diff.ID, headerID, ttl)
-	})
-
-	It("reads in a kicks storage diff and persists it", func() {
-		//TODO: update this when we get a storage diff row for Flop kicks
-	})
-
-	It("reads in a live storage diff and persists it", func() {
-		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000008")
-		value := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000001")
-		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
-		err := transformer.Execute(diff)
-		Expect(err).NotTo(HaveOccurred())
-
-		var liveResult test_helpers.VariableRes
-		err = db.Get(&liveResult, `SELECT diff_id, header_id, live AS value from maker.flop_live`)
-		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(liveResult, diff.ID, headerID, "1")
-	})
-
-	It("reads in a vow storage diff and persists it", func() {
-		vow := "0x1CC5ABe5C0464F3af2a10df0c711236a8446BF75"
-		key := common.HexToHash("0000000000000000000000000000000000000000000000000000000000000009")
-		value := common.HexToHash("0000000000000000000000001cc5abe5c0464f3af2a10df0c711236a8446bf75")
-		diff := test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
 
 		err := transformer.Execute(diff)
 		Expect(err).NotTo(HaveOccurred())
 
-		var vowResult test_helpers.VariableRes
-		err = db.Get(&vowResult, `SELECT diff_id, header_id, vow AS value from maker.flop_vow`)
+		var ttlResult test_helpers.VariableRes
+		err = db.Get(&ttlResult,
+			`SELECT diff_id, header_id, tau AS value FROM maker.flip_tau`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(vowResult, diff.ID, headerID, vow)
+		test_helpers.AssertVariable(ttlResult, diff.ID, header.Id, "172800")
+	})
+
+	XIt("reads in a kicks storage diff and persists it", func() {
+		//TODO: update this when we get a storage diff row for Flap kicks
 	})
 
 	Describe("wards", func() {
@@ -184,8 +147,8 @@ var _ = Describe("Executing the flop transformer", func() {
 			denyLog := test_data.CreateTestLog(header.Id, db)
 			denyModel := test_data.DenyModel()
 
-			flopAddressID, flopAddressErr := shared.GetOrCreateAddress(test_data.FlopAddress(), db)
-			Expect(flopAddressErr).NotTo(HaveOccurred())
+			flipAddressID, flipAddressErr := shared.GetOrCreateAddress(test_data.EthFlipAddress(), db)
+			Expect(flipAddressErr).NotTo(HaveOccurred())
 
 			userAddress := "0xffb0382ca7cfdc4fc4d5cc8913af1393d7ee1ef1"
 			userAddressID, userAddressErr := shared.GetOrCreateAddress(userAddress, db)
@@ -197,7 +160,7 @@ var _ = Describe("Executing the flop transformer", func() {
 
 			denyModel.ColumnValues[event.HeaderFK] = header.Id
 			denyModel.ColumnValues[event.LogFK] = denyLog.ID
-			denyModel.ColumnValues[event.AddressFK] = flopAddressID
+			denyModel.ColumnValues[event.AddressFK] = flipAddressID
 			denyModel.ColumnValues[constants.MsgSenderColumn] = msgSenderAddressID
 			denyModel.ColumnValues[constants.UsrColumn] = userAddressID
 			insertErr := event.PersistModels([]event.InsertionModel{denyModel}, db)
@@ -213,13 +176,13 @@ var _ = Describe("Executing the flop transformer", func() {
 			var wardsResult test_helpers.WardsMappingRes
 			err := db.Get(&wardsResult, `SELECT diff_id, header_id, address_id, usr AS key, wards.wards AS value FROM maker.wards`)
 			Expect(err).NotTo(HaveOccurred())
-			Expect(wardsResult.AddressID).To(Equal(strconv.FormatInt(flopAddressID, 10)))
+			Expect(wardsResult.AddressID).To(Equal(strconv.FormatInt(flipAddressID, 10)))
 			test_helpers.AssertMapping(wardsResult.MappingRes, wardsDiff.ID, header.Id, strconv.FormatInt(userAddressID, 10), "1")
 		})
 	})
 
 	Describe("bids", func() {
-		//TODO: update when we get real flop bid storage diffs
+		//TODO: update when we get real flip bid storage diffs
 		Describe("guy + tic + end packed slot", func() {
 			var (
 				bidId int
@@ -232,10 +195,10 @@ var _ = Describe("Executing the flop transformer", func() {
 				value := common.HexToHash("00000002a300000000002a30284ecb5880cdc3362d979d07d162bf1d8488975d")
 				diff = test_helpers.CreateDiffRecord(db, header, transformer.HashedAddress, key, value)
 
-				addressId, addressErr := shared.GetOrCreateAddress(flopperContractAddress, db)
+				addressId, addressErr := shared.GetOrCreateAddress(contractAddress, db)
 				Expect(addressErr).NotTo(HaveOccurred())
 
-				_, writeErr := db.Exec(flop.InsertFlopKicksQuery, diff.ID, headerID, addressId, bidId)
+				_, writeErr := db.Exec(flip.InsertFlipKicksQuery, diff.ID, header.Id, addressId, bidId)
 				Expect(writeErr).NotTo(HaveOccurred())
 
 				executeErr := transformer.Execute(diff)
@@ -244,23 +207,23 @@ var _ = Describe("Executing the flop transformer", func() {
 
 			It("reads and persists a guy diff", func() {
 				var bidGuyResult test_helpers.MappingRes
-				dbErr := db.Get(&bidGuyResult, `SELECT diff_id, header_id, bid_id AS key, guy AS value FROM maker.flop_bid_guy`)
+				dbErr := db.Get(&bidGuyResult, `SELECT diff_id, header_id, bid_id AS key, guy AS value FROM maker.flip_bid_guy`)
 				Expect(dbErr).NotTo(HaveOccurred())
-				test_helpers.AssertMapping(bidGuyResult, diff.ID, headerID, strconv.Itoa(bidId), "0x284ecB5880CdC3362D979D07D162bf1d8488975D")
+				test_helpers.AssertMapping(bidGuyResult, diff.ID, header.Id, strconv.Itoa(bidId), "0x284ecB5880CdC3362D979D07D162bf1d8488975D")
 			})
 
 			It("reads and persists a tic diff", func() {
 				var bidTicResult test_helpers.MappingRes
-				dbErr := db.Get(&bidTicResult, `SELECT diff_id, header_id, bid_id AS key, tic AS value FROM maker.flop_bid_tic`)
+				dbErr := db.Get(&bidTicResult, `SELECT diff_id, header_id, bid_id AS key, tic AS value FROM maker.flip_bid_tic`)
 				Expect(dbErr).NotTo(HaveOccurred())
-				test_helpers.AssertMapping(bidTicResult, diff.ID, headerID, strconv.Itoa(bidId), "10800")
+				test_helpers.AssertMapping(bidTicResult, diff.ID, header.Id, strconv.Itoa(bidId), "10800")
 			})
 
 			It("reads and persists an end diff", func() {
 				var bidEndResult test_helpers.MappingRes
-				dbErr := db.Get(&bidEndResult, `SELECT diff_id, header_id, bid_id AS key, "end" AS value FROM maker.flop_bid_end`)
+				dbErr := db.Get(&bidEndResult, `SELECT diff_id, header_id, bid_id AS key, "end" AS value FROM maker.flip_bid_end`)
 				Expect(dbErr).NotTo(HaveOccurred())
-				test_helpers.AssertMapping(bidEndResult, diff.ID, headerID, strconv.Itoa(bidId), "172800")
+				test_helpers.AssertMapping(bidEndResult, diff.ID, header.Id, strconv.Itoa(bidId), "172800")
 			})
 		})
 	})
