+++ transformers/component_tests/jug/configured_transformer_test.go
@@ -47,23 +47,21 @@ var _ = Describe("Executing the transformer", func() {
 			StorageKeysLookup: storageKeysLookup,
 			Repository:        &repository,
 		}
-		header   = fakes.FakeHeader
-		headerID int64
-		err      error
-		ilkID    int64
+		header = fakes.FakeHeader
+		ilkID  int64
 	)
 
 	BeforeEach(func() {
 		test_config.CleanTestDB(db)
 		transformer.NewTransformer(db)
 		ilk := "0x4554480000000000000000000000000000000000000000000000000000000000"
-		ilkID, err = shared.GetOrCreateIlk(ilk, db)
-		Expect(err).NotTo(HaveOccurred())
+		var ilkErr error
+		ilkID, ilkErr = shared.GetOrCreateIlk(ilk, db)
+		Expect(ilkErr).NotTo(HaveOccurred())
 		headerRepository := repositories.NewHeaderRepository(db)
 		var insertHeaderErr error
-		headerID, insertHeaderErr = headerRepository.CreateOrUpdateHeader(header)
+		header.Id, insertHeaderErr = headerRepository.CreateOrUpdateHeader(header)
 		Expect(insertHeaderErr).NotTo(HaveOccurred())
-		header.Id = headerID
 	})
 
 	It("reads in a Jug Vat storage diff row and persists it", func() {
@@ -77,7 +75,7 @@ var _ = Describe("Executing the transformer", func() {
 		var vatResult test_helpers.VariableRes
 		err = db.Get(&vatResult, `SELECT diff_id, header_id, vat AS value FROM maker.jug_vat`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(vatResult, jugVatDiff.ID, headerID, "0x67fd6c3575Fc2dBE2CB596bD3bEbc9EDb5571fA1")
+		test_helpers.AssertVariable(vatResult, jugVatDiff.ID, header.Id, "0x67fd6c3575Fc2dBE2CB596bD3bEbc9EDb5571fA1")
 	})
 
 	It("reads in a Jug Vow storage diff row and persists it", func() {
@@ -91,7 +89,7 @@ var _ = Describe("Executing the transformer", func() {
 		var vowResult test_helpers.VariableRes
 		err = db.Get(&vowResult, `SELECT diff_id, header_id, vow AS value FROM maker.jug_vow`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertVariable(vowResult, jugVowDiff.ID, headerID, "0x17560834075da3db54f737db74377e799c865821000000000000000000000000")
+		test_helpers.AssertVariable(vowResult, jugVowDiff.ID, header.Id, "0x17560834075da3db54f737db74377e799c865821000000000000000000000000")
 	})
 
 	It("reads in a wards storage diff row and persists it", func() {
@@ -142,7 +140,7 @@ var _ = Describe("Executing the transformer", func() {
 		var ilkDutyResult test_helpers.MappingRes
 		err = db.Get(&ilkDutyResult, `SELECT diff_id, header_id, ilk_id AS key, duty AS value FROM maker.jug_ilk_duty`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertMapping(ilkDutyResult, jugIlkDutyDiff.ID, headerID, strconv.FormatInt(ilkID, 10), "1000000000000000000000000000")
+		test_helpers.AssertMapping(ilkDutyResult, jugIlkDutyDiff.ID, header.Id, strconv.FormatInt(ilkID, 10), "1000000000000000000000000000")
 	})
 
 	It("reads in a Jug Ilk Rho storage diff row and persists it", func() {
@@ -156,6 +154,6 @@ var _ = Describe("Executing the transformer", func() {
 		var ilkRhoResult test_helpers.MappingRes
 		err = db.Get(&ilkRhoResult, `SELECT diff_id, header_id, ilk_id AS key, rho AS value FROM maker.jug_ilk_rho`)
 		Expect(err).NotTo(HaveOccurred())
-		test_helpers.AssertMapping(ilkRhoResult, jugIlkRhoDiff.ID, headerID, strconv.FormatInt(ilkID, 10), "1551968264")
+		test_helpers.AssertMapping(ilkRhoResult, jugIlkRhoDiff.ID, header.Id, strconv.FormatInt(ilkID, 10), "1551968264")
 	})
 })
