+++ transformers/component_tests/pot/configured_transformer_test.go
