+++ transformers/component_tests/spot/configured_transformer_test.go
